Fix infinite loop in do_sieve.
There was an infinite loop in do_sieve caused by a call to std::fill_n
with a segment size of 0, which happened when offset == limit. This
would be the case when sieve_limit was called with a limit equal to
primes.back() + 2. In that case only primes.back() + 2 should be
checked for primality and added if appropriate.